requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
bcrypt==4.2.1
binaryornot==0.4.4
boto3==1.35.78
botocore==1.35.78
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.19
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
